upate

+++ SFML_RPG/src/components/AttributeComponent.cpp
@@ -1,11 +1,11 @@
 #include "stdafx.h"
 #include "AttributeComponent.h"
 
-AttributeComponent::AttributeComponent(unsigned level)
+AttributeComponent::AttributeComponent(int level)
 {
 	m_Level = level;
 	m_Exp = 0;
-	m_ExpNext = static_cast<unsigned>((50 / 3) * (pow(m_Level + 1, 3) - 6 * pow(m_Level + 1, 2) + ((m_Level + 1) * 17) - 12));
+	m_ExpNext = (50 / 3) * static_cast<int>(pow(m_Level + 1, 3) - 6 * static_cast<int>(pow(m_Level + 1, 2)) + ((m_Level + 1) * 17) - 12);
 	m_AttributePoints = 2;
 
 	m_Vitality = 1;
@@ -44,12 +44,12 @@ void AttributeComponent::updateLevel()
 	{
 		++m_Level;
 		m_Exp -= m_ExpNext;
-		m_ExpNext = static_cast<unsigned>((50 / 3) * (pow(m_Level, 3) - 6 * pow(m_Level, 2) + (m_Level * 17) - 12));
+		m_ExpNext = (50 / 3) * static_cast<int>(pow(m_Level, 3) - 6 * pow(m_Level, 2) + (m_Level * 17) - 12);
 		++m_AttributePoints;
 	}
 }
 
-void AttributeComponent::gainExp(const unsigned exp)
+void AttributeComponent::gainExp(const int exp)
 {
 	m_Exp += exp;
